requirements.txt
streamlit>=1.44,<2.0
pandas>=2.2,<3.0
sentence-transformers>=3.0,<4.0
faiss-cpu>=1.8,<2.0
